+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Statistics grouped rows: sub CQP only has one of the values [#452](https://github.com/spraakbanken/korp-frontend/issues/452)
+- Decimal separator not updated in statistics when changing language [#246](https://github.com/spraakbanken/korp-frontend/issues/246)
 
 ## [9.9.0] - 2025-03-24